GET,POST

+++ src/main/java/com/java/salary_income/application/service/CalculateService.java
@@ -16,7 +16,7 @@ public class CalculateService {
         this.calculateRepository = calculateRepository;
     }
 
-    public float getYearlySalary() {
+    public int getYearlySalary() {
         return calculateRepository.getYearlySalary();
     }
