+++ src/main/java/com/java/salary_income/domain/repository/CalculateRepository.java
@@ -4,7 +4,7 @@ import com.java.salary_income.domain.entity.SalaryInfoEntity;
 
 public interface CalculateRepository {
 
-    float getYearlySalary();
+    int getYearlySalary();
 
     void post(SalaryInfoEntity salaryInfoEntity);
 }
